app/requirements.txt
mysql-connector-python
flask
os